Set Contact as default controller and update navigation
Changed the default route in `Program.cs` to use `ContactController` as the default controller. Updated `_Layout.cshtml` to simplify navigation by removing "Home" and "Privacy" links from the navbar and footer. Added `ContactController` with an `Index` action to display a list of contacts. Created a new `Index.cshtml` view for listing contacts with a "Create" button.